+++ excellent/types/time_test.go
@@ -12,21 +12,19 @@ import (
 )
 
 func TestXTime(t *testing.T) {
-	// test stringing
-	assert.Equal(t, `17:01:30.000000`, types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0)).String())
-
-	// test equality
-	assert.True(t, types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0)).Equals(types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0))))
-	assert.False(t, types.NewXDateTime(time.Date(2019, 4, 9, 17, 1, 30, 0, time.UTC)).Equals(types.NewXDateTime(time.Date(2018, 4, 9, 17, 1, 30, 0, time.UTC))))
-
-	// test comparisons
-	assert.Equal(t, 0, types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0)).Compare(types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0))))
-	assert.Equal(t, 1, types.NewXTime(utils.NewTimeOfDay(17, 1, 31, 0)).Compare(types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0))))
-	assert.Equal(t, -1, types.NewXTime(utils.NewTimeOfDay(17, 1, 29, 0)).Compare(types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0))))
-
 	t1 := types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0))
 	assert.Equal(t, t1, t1.Reduce(utils.NewEnvironmentBuilder().Build()))
 	assert.Equal(t, `time`, t1.Describe())
+	assert.Equal(t, `17:01:30.000000`, types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0)).String())
+
+	// test equality
+	assert.True(t, t1.Equals(types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0))))
+	assert.False(t, t1.Equals(types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 1))))
+
+	// test comparisons
+	assert.Equal(t, 0, types.NewXTime(utils.NewTimeOfDay(17, 1, 30, 0)).Compare(t1))
+	assert.Equal(t, 1, types.NewXTime(utils.NewTimeOfDay(17, 1, 31, 0)).Compare(t1))
+	assert.Equal(t, -1, types.NewXTime(utils.NewTimeOfDay(17, 1, 29, 0)).Compare(t1))
 }
 
 func TestToXTime(t *testing.T) {
@@ -40,9 +38,8 @@ func TestToXTime(t *testing.T) {
 		{types.NewXNumberFromInt(123), types.XTimeZero, true},
 		{types.NewXText("10:30"), types.NewXTime(utils.NewTimeOfDay(10, 30, 0, 0)), false},
 		{types.NewXText("10:30 pm"), types.NewXTime(utils.NewTimeOfDay(22, 30, 0, 0)), false},
-		// TODO hour only parsing
-		//{types.NewXText("10"), types.NewXTime(utils.NewTimeOfDay(10, 0, 0, 0)), false},
-		//{types.NewXText("10 PM"), types.NewXTime(utils.NewTimeOfDay(22, 0, 0, 0)), false},
+		{types.NewXText("10"), types.NewXTime(utils.NewTimeOfDay(10, 0, 0, 0)), false},
+		{types.NewXText("10 PM"), types.NewXTime(utils.NewTimeOfDay(22, 0, 0, 0)), false},
 		{types.NewXText("wha?"), types.XTimeZero, true},
 		{NewTestXObject("Hello", 123), types.XTimeZero, true},
 		{NewTestXObject("10:30:24", 123), types.NewXTime(utils.NewTimeOfDay(10, 30, 24, 0)), false},